+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Models/MobileTextureFormat.cs
@@ -8,16 +8,17 @@ namespace KRT.VRCQuestTools.Models
     public enum MobileTextureFormat
     {
 #pragma warning disable SA1602 // Disable SA1602 for enum documentation
-        /// <summary>
-        /// Don't set platform-specific override for texture import settings.
-        /// </summary>
-        NoOverride = -1,
         ASTC_4x4 = TextureFormat.ASTC_4x4,
         ASTC_5x5 = TextureFormat.ASTC_5x5,
         ASTC_6x6 = TextureFormat.ASTC_6x6,
         ASTC_8x8 = TextureFormat.ASTC_8x8,
         ASTC_10x10 = TextureFormat.ASTC_10x10,
         ASTC_12x12 = TextureFormat.ASTC_12x12,
+
+        /// <summary>
+        /// Don't set platform-specific override for texture import settings.
+        /// </summary>
+        NoOverride = -1,
 #pragma warning restore SA1602
     }
 }